+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - OAuth2client: support CredentialsRepo
 - remove dependency on abstract-sdk
+- extend AbortError(requestOptions)
 
 ### Deprecated